+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -19,6 +19,7 @@
 - Files with extension should not be set to hlasm in libs folder
 - Lookahead mode does not work correctly when triggered from AINSERTed code
 - Incorrect relative immediate operand validation
+- Remove ALIAS operand parsing limitation
 
 ## [0.14.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/0.13.0...0.14.0) (2021-08-18)